Fixed the problem of not being able to use decimals by changing int to float

+++ 02_Ingredients/Ingredients_function_V2.py
@@ -13,7 +13,7 @@ def item_cost(question):
 	valid = False
 	while not valid:
 		try:
-			response = int(input(question))
+			response = float(input(question))
 			if response <= 0:
 				print(error)
 			else:
